split into server component wrapper and client component
